:sparkles: (Cheatcodes) Create error navigation
